+++ lib/src/eval/bindgen/type.dart
@@ -2,8 +2,12 @@ import 'package:analyzer/dart/element/element.dart';
 import 'package:analyzer/dart/element/nullability_suffix.dart';
 import 'package:analyzer/dart/element/type.dart';
 import 'package:dart_eval/src/eval/bindgen/context.dart';
+import 'package:dart_eval/src/eval/bindgen/errors.dart';
 
 String bridgeTypeRefFromType(DartType type) {
+  if (type is TypeParameterType) {
+    return 'BridgeTypeRef.ref(\'${type.element.name}\')';
+  }
   return 'BridgeTypeRef(${bridgeTypeSpecFrom(type)})';
 }
 
@@ -29,6 +33,9 @@ String? builtinTypeFrom(DartType type) {
   }
   if (type is DynamicType) {
     return 'CoreTypes.dynamicType';
+  }
+  if (type is FunctionType) {
+    return 'CoreTypes.function';
   }
 
   final element = type.element!;
@@ -66,8 +73,12 @@ String? builtinTypeFrom(DartType type) {
   return null;
 }
 
-String? wrapVar(BindgenContext ctx, DartType type, String expr) {
+String? wrapVar(BindgenContext ctx, DartType type, String expr,
+    [bool forFunc = false]) {
   if (type is VoidType) {
+    if (forFunc) {
+      return '\$null()';
+    }
     return 'null';
   }
 
@@ -80,9 +91,9 @@ String? wrapVar(BindgenContext ctx, DartType type, String expr) {
   if (wrapped == null) {
     if (ctx.unknownTypes.add(type.element!.name!)) {
       print('Warning: type ${type.element!.name} is not bound, '
-          'falling back to \$Object');
+          'falling back to wrapAlways()');
     }
-    wrapped = '\$Object($expr)';
+    wrapped = 'runtime.wrapAlways($expr)';
   }
 
   if (type.nullabilitySuffix == NullabilitySuffix.question) {
@@ -101,7 +112,20 @@ String? wrapType(BindgenContext ctx, DartType type, String expr) {
     return '\$null()';
   }
 
-  final element = type.element!;
+  if (type is DynamicType) {
+    return '\$Object($expr)';
+  }
+
+  if (type is FunctionType) {
+    return wrapFunctionType(ctx, type, expr);
+  }
+
+  if (type.isDartCoreFunction) {
+    return '\$Function((runtime, target, args) => $expr())';
+  }
+
+  final element = type.element ??
+      (throw BindingGenerationError('Type $type has no element'));
   final lib = element.library!;
   final name = element.name ?? ' ';
 
@@ -129,5 +153,75 @@ String? wrapType(BindgenContext ctx, DartType type, String expr) {
     return '\$$name.wrap($expr)';
   }
 
+  if (type is TypeParameterType) {
+    final bound = type.bound;
+    if (bound is! DynamicType) {
+      return wrapVar(ctx, bound, expr);
+    }
+  }
+
   return null;
+}
+
+String wrapFunctionType(BindgenContext ctx, FunctionType type, String expr) {
+  var buffer = StringBuffer('\$Function((runtime, target, args) { ');
+  if (type.returnType is! VoidType && !type.returnType.isDartCoreNull) {
+    buffer.write('final funcResult = ');
+  }
+  buffer.write('$expr(');
+  var i = 0;
+  for (; i < type.normalParameterTypes.length; i++) {
+    buffer.write('args[$i]');
+    final _type = type.normalParameterTypes[i];
+    if (_type.nullabilitySuffix == NullabilitySuffix.question) {
+      buffer.write('?.\$value');
+    } else {
+      buffer.write('!.\$value');
+    }
+    if (i < type.normalParameterTypes.length - 1) {
+      buffer.write(', ');
+    }
+  }
+
+  if (type.optionalParameterNames.isNotEmpty) {
+    for (var j = i; j < type.optionalParameterNames.length + i; j++) {
+      if (type.normalParameterTypes.isNotEmpty) {
+        buffer.write(', ');
+      }
+      final _type = type.optionalParameterTypes[i];
+      buffer.write('args[$j]');
+      if (_type.nullabilitySuffix == NullabilitySuffix.question) {
+        buffer.write('?.\$value');
+      } else {
+        buffer.write('!.\$value');
+      }
+      if (j < type.optionalParameterNames.length + i - 1) {
+        buffer.write(', ');
+      }
+    }
+  }
+
+  if (type.namedParameterTypes.isNotEmpty) {
+    if (type.normalParameterTypes.isNotEmpty ||
+        type.optionalParameterNames.isNotEmpty) {
+      buffer.write(', ');
+    }
+
+    var k = i;
+    type.namedParameterTypes.forEach((_name, _type) {
+      buffer.write(_name);
+      buffer.write(': args[$k]');
+      if (_type.nullabilitySuffix == NullabilitySuffix.question) {
+        buffer.write('?.\$value');
+      } else {
+        buffer.write('!.\$value');
+      }
+      if (k < type.namedParameterTypes.length + i - 1) {
+        buffer.write(', ');
+      }
+    });
+  }
+  buffer.write(
+      '); return ${wrapVar(ctx, type.returnType, 'funcResult', true)}; })');
+  return buffer.toString();
 }