+++ lib/src/eval/shared/stdlib/io.dart
@@ -27,6 +27,10 @@ class DartIoPlugin implements EvalPlugin {
     $InternetAddress.configureForCompile(registry);
     $InternetAddressType.configureForCompile(registry);
     registry.addSource($HttpStatusSource());
+    registry.addSource(DartSource('dart:io', '''
+    library dart.io;
+    export 'dart:io/http_status.dart';
+    '''));
   }
 
   @override